pyproject.toml
[project]
name = "qspr-explainability"
version = "0.1.0"
description = "Explainable QSAR and QSPR experiments with fingerprint-based models and motif visualization."
requires-python = ">=3.11"
dependencies = [
    "BorutaShap==1.0.17",
    "ipykernel==6.30.1",
    "jupyterlab==4.4.7",
    "matplotlib==3.10.6",
    "nbconvert==7.16.6",
    "nbformat==5.10.4",
    "numba==0.62.1",
    "numpy==1.26.4",
    "pandas==2.3.2",
    "rdkit-pypi==2022.9.5",
    "scikit-learn==1.7.2",
    "scipy==1.16.3",
    "seaborn==0.13.2",
    "shap==0.49.1",
    "statsmodels==0.14.6",
    "xgboost==1.7.6",
    "iterative-stratification==0.1.9",
    "pillow==11.2.1",
]
